api-backend/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.5.2
python-dotenv==1.0.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# MongoDB dependencies  
motor==3.3.2
pymongo==4.6.1

# Additional dependencies
python-multipart==0.0.6
httpx==0.25.2
email-validator==2.1.0

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
